refactor: simplify structure
