+++ .github/workflows/pre-merge.yaml
@@ -23,5 +23,7 @@ jobs:
         uses: actions/checkout@8ade135a41bc03ea155e62e844d188df1ea18608 # https://github.com/actions/checkout
       - name: Cache Gradle Caches
         uses: gradle/gradle-build-action@v2 # https://github.com/gradle/gradle-build-action
+      - name: Make gradlew executable
+        run: chmod +x ./gradlew
       - name: Run Gradle tasks
         run: ./gradlew preMerge --continue
