--- example/src/main/resources/avro/Example.avsc
@@ -1,23 +1,0 @@
-{
-    "type": "record",
-    "namespace": "Example",
-    "name": "Example",
-    "fields": [
-        {
-            "name": "Name",
-            "type": "string"
-        },
-        {
-            "name": "Age",
-            "type": "int"
-        },
-        {
-            "name": "userName",
-            "type": [
-                "null",
-                "string"
-            ],
-            "default": null
-        }
-    ]
-}
