--- example/src/main/resources/avro/ExampleProtocol.avpr
@@ -1,30 +1,0 @@
-{
-  "protocol" : "ExampleProtocol",
-  "namespace" : "ru.rudikov.example",
-  "types" : [ {
-    "type" : "record",
-    "name" : "FirstExampleRecordFromProtocol",
-    "fields" : [ {
-      "name" : "name",
-      "type" : "string"
-    }, {
-      "name" : "age",
-      "type" : "int"
-    }, {
-      "name" : "userName",
-      "type" : [ "null", "string" ],
-      "default" : null
-    } ]
-  }, {
-    "type" : "record",
-    "name" : "SecondExampleRecordFromProtocol",
-    "fields" : [ {
-      "name" : "name",
-      "type" : "string"
-    }, {
-      "name" : "color",
-      "type" : "string"
-    } ]
-  } ],
-  "messages" : { }
-}
