+++ gradle.properties
@@ -1,3 +1,12 @@
+ID=io.github.rudikone.avroschema-wizard-plugin
+VERSION=3.0.0
+GROUP=io.github.rudikone
+DISPLAY_NAME=Avro Schema Wizard Gradle Plugin
+DESCRIPTION=Gradle Plugin for working with Avro schemas in Schema Registry
+WEBSITE=https://github.com/rudikone/avroschema-wizard-plugin
+VCS_URL=https://github.com/rudikone/avroschema-wizard-plugin
+IMPLEMENTATION_CLASS=io.github.rudikone.avroschemawizardplugin.AvroSchemaWizardPlugin
+
 org.gradle.jvmargs=-Xmx1536m
 org.gradle.parallel=true
