+++ gradle/libs.versions.toml
@@ -1,6 +1,5 @@
 [versions]
 detekt = "1.23.4"
-kotlin = "1.9.22"
 ktlintGradle = "12.0.3"
 pluginPublish = "1.2.1"
 versionCheck = "0.50.0"
@@ -11,7 +10,6 @@ shadow = "8.1.1"
 
 [plugins]
 detekt = { id = "io.gitlab.arturbosch.detekt", version.ref = "detekt"}
-kotlin = { id = "org.jetbrains.kotlin.jvm", version.ref = "kotlin"}
 ktlint = { id = "org.jlleitschuh.gradle.ktlint", version.ref = "ktlintGradle"}
 pluginPublish = { id = "com.gradle.plugin-publish", version.ref = "pluginPublish"}
 versionCheck = { id = "com.github.ben-manes.versions", version.ref = "versionCheck"}
@@ -20,4 +18,4 @@ shadow = { id = "com.github.johnrengelman.shadow", version.ref = "shadow"}
 [libraries]
 junit = { module = "junit:junit", version.ref = "junit" }
 schemaRegistryClient = { module = "io.confluent:kafka-schema-registry-client", version.ref = "schemaRegistryClient"}
-kafkaSchemaSerializer = { module = "io.confluent:kafka-schema-serializer", version.ref = "schemaRegistryClient"}
+kafkaSchemaSerializer = { module = "io.confluent:kafka-schema-serializer", version.ref = "kafkaSchemaSerializer"}
