--- plugin-build/gradle.properties
@@ -1,8 +1,0 @@
-ID=io.github.rudikone.avroschema-wizard-plugin
-VERSION=2.0.1
-GROUP=io.github.rudikone
-DISPLAY_NAME=Avro Schema Wizard Gradle Plugin
-DESCRIPTION=Gradle Plugin for working with Avro schemas in Schema Registry
-WEBSITE=https://github.com/rudikone/avroschema-wizard-plugin
-VCS_URL=https://github.com/rudikone/avroschema-wizard-plugin
-IMPLEMENTATION_CLASS=io.github.rudikone.plugin.AvroSchemaWizardPlugin
