+++ src/main/java/io/github/rudikone/avroschemawizardplugin/CompatibilityCheckTask.kt
@@ -1,4 +1,4 @@
-package io.github.rudikone.plugin
+package io.github.rudikone.avroschemawizardplugin
 
 import io.confluent.kafka.schemaregistry.avro.AvroSchema
 import io.confluent.kafka.schemaregistry.client.CachedSchemaRegistryClient
@@ -83,7 +83,7 @@ abstract class CompatibilityCheckTask : DefaultTask() {
             SubjectNameStrategies.from(config.subjectNameStrategy.get())
                 ?: error(
                     "Unsupported subject name strategy! " +
-                        "Allowed values: ${SubjectNameStrategies.entries.toTypedArray()}",
+                        "Allowed values: ${SubjectNameStrategies.values()}",
                 )
 
         val nameStrategy =
